bugfix: fix some problems with zstd compressor
