+++ compressor/seata-compressor-zstd/src/main/resources/META-INF/services/io.seata.core.compressor.Compressor
@@ -1,1 +1,1 @@
-io.seata.compressor.zip.ZstdCompressor
+io.seata.compressor.zstd.ZstdCompressor
